Gates part for authorization

+++ app/Http/Controllers/QuestionsController.php
@@ -6,12 +6,13 @@ use App\Http\Requests\Questions\CreateQuestionRequest;
 use App\Http\Requests\Questions\UpdateQuestionRequest;
 use App\Question;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionsController extends Controller
 {
    
     public function ___construct(){
-        $this->middleware(['auth'])->only(['create','store']);
+        $this->middleware(['auth'])->only(['create','store','edit','update']);
     }
     public function index()
     {
@@ -43,10 +44,18 @@ class QuestionsController extends Controller
         return redirect(route('questions.index'));
     }
 
-    
-    public function show(Question $question)
-    {
-        //
+    // METHOD 1    
+    // public function show($slug)
+    // {
+    //     $question = Question::where('slug', $slug)->firstOrFail();
+    // }
+
+    // METHOD 2
+    public function show(Question $question){
+        $question->increment('views_count');//this method will increment the count of var899
+        return view('questions.show', compact([
+            'question'
+        ]));
     }
 
     public function edit(Question $question)
@@ -60,19 +69,45 @@ class QuestionsController extends Controller
     
     public function update(UpdateQuestionRequest $request, Question $question)
     {
-        $question->update([
-            'title'=>$request->title,
-            'body'=>$request->body
-        ]);
-        session()->flash('success', 'Question has been updated SuccessFully !');
-        return redirect(route('questions.index'));
+        //method 1 for calling gate
+        if(Gate::allows('update-question', $question)){
+            $question->update([
+                'title'=>$request->title,
+                'body'=>$request->body
+            ]);
+            session()->flash('success', 'Question has been updated SuccessFully !');
+            return redirect(route('questions.index'));
+        }
+        abort(403);
     }
 
     
     public function destroy(Question $question)
     {
-        $question->delete();
-        session()->flash('success','Question has been added deleted Successfully!');
-        return redirect(route('questions.index'));
-    }
+        //method 2 for calling gate
+        if(auth()->user()->can('delete-question', $question)){
+            $question->delete();
+            session()->flash('success','Question has been added deleted Successfully!');
+            return redirect(route('questions.index'));
+        }
+        abort(403);
+    }    
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
